refactor(App.js) : function 폴더를 만들어, 자동차 경기/결과/예외 코드 나누기
